Use dynamic gradle dependency
Yeah, I know. Using dynamic dependency can cause unexpected version updates and difficulty resolving version differences. But I want to use vector drawables and the gradle plugin v1.2.3 doesn't support that. Tried to update it to 2.3.3 but faced a lot of errors, it just doesn't work with this project. The dependencies and libraries used in the project are very old and not supported in newer versions of gradle.